Refuse DROP COMPACT STORAGE if some 2.x sstables are in use
The 'DROP COMPACT STORAGE' functionality rely on the fact that starting
with 3.0, compact storage tables have no special layout internally: they
are like non-compact tables but the CQL layer simply impose restriction
on them (for backward compatibility). As such, 'DROP COMPACT STORAGE'
simply instruct CQL to stop imposing those restrictions, but nothing
happens to the internal (sstable level) table representation.

This is not true in 2.x: in the 2.x internal format, a compact table
will have a different internal representation than its non-compact
counterpart. As such, the sstables of a compact table cannot be properly
read by a 3.x node unless it knows the table is compact.

Tl;dr, if DROP COMPACT STORAGE is called on a table for which there is
still non-upgraded 2.x sstables in the cluster, then those sstables
won't be readable anymore.

This patch ensures that DROP COMPACT STORAGE is not allowed until all
sstables have been upgraded to the 3.x format, thus ensuring it is safe
to call. Note that for DROP COMPACT STORAGE to be safe, we need to
ensure there is no more 2.x sstables _anywhere_ in the cluster. So this
patch adds the gossiping of the current sstable versions in use by each
node in order to implement its check.